requirements.txt
wheel
urllib3
Django>=2
Pillow

# for mac address field
netaddr

# for testing required xmltodict
#xmltodict
dicttoxml

# db client for Mariadb
mysqlclient

-e git://github.com/nerosketch/easysnmp.git#egg=easysnmp
pid
django-guardian
pinax-theme-bootstrap
django-bootstrap3

# django-jsonfield
-e git://github.com/dmkoch/django-jsonfield.git#egg=django-jsonfield

requests
webdavclient
transliterate
django-encrypted-model-fields

# For compatible ip_pool app
django-netfields

# django-xmlview for pay system allpay
-e git://github.com/nerosketch/django-xmlview.git#egg=django-xmlview

# django-bitfield
-e git://github.com/disqus/django-bitfield.git#egg=django-bitfield

# django_cleanup for clean unused media
-e git://github.com/un1t/django-cleanup.git#egg=django_cleanup

# viberbot
-e git://github.com/Viber/viber-bot-python.git#egg=viberbot

# pexpect
-e git://github.com/pexpect/pexpect.git#egg=pexpect

Celery
redis
celery[redis]

docxtpl
